MB-58863: Permit certain swap rebalances, ignoring data GRs
We currently check rebalance GuardRails whenever a live node is being
ejected. This prevents perfectly safe rebalances such as a swap
rebalance. Instead we should only make the checks when the number of
kv nodes will be strictly less after the rebalance.

To ensure that we don't permit unsafe bucket migrations to continue,
we will still make the check for a bucket if a storage migration is in
progress, even if the number of kv nodes has reduced.
This ensures that in a magma to couchstore conversion, we halt the
migration once the RR% is too low for couchstore (<1% by default
currently). For consistency with the initial guardrail check fo
migration, we also halt a couchstore to magma migration if the RR% is
too low for couchstore.

Change-Id: I4b5f01b1eb140dd841deea6502cef610c9809dc2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198148
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>