MB-58931: Enforce data size guardrail on rebalance
Enforce the data size per node maximum on rebalances, with the same
logic as we use for resident ratio. That is, a rebalance is rejected if
the number of live nodes in the cluster is going down, and the data
size per node is expected to violate the guardrail when performing the
rebalance.

Change-Id: Ic0f43487e66e8d73841d79b979c7a004d0a6f2e1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198371
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>