[cluster_tests] Rebalance should wait for ejected nodes
This is a correction for
https://review.couchbase.org/c/ns_server/+/197706

It is not correct to wait for a web server start at ejected nodes
because when the rebalance has finished the web server might still
be running. We can wait for GET /pools/default to return 404 instead.
This means the node is up and the node is uninitialized (which means
it has left the cluster).

Change-Id: I8569a3f7d06787c26d10cef7d6bfb2f5100ee306
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198002
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>