Increase default print depth of eunit_tty from 20 to 100.
If assert macros fail and the terms provided to them are too long then
parts of the term will be replaced with '...'. This makes debugging more
difficult as one would otherwise have to print the terms manually to see
their full extent.

Bump the print depth provided to eunit_tty (via eunit:test/2) to print
longer terms by default when assert macros fail and print (this does not
affect things logged via ?log macros). The value 100 is arbitrary, but
was observed to print vBucket maps and collections manifests that were
otherwise not printed. It can easily be changed in the future if
required.

Change-Id: Idc3b986a562167542a16fe8e902ebe5e6a39113b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198029
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>