MB-58898: Hold quorum and sync config before auto-failover checks
Consider a scenario in which we have a network partition with as
follows:

    Partition A - [a, b, q]
    Partition B - [c, d]

In which [a, b, c, d] are KV nodes, and q runs any other service to
provide one side of the partition with a viable quorum.

N.B. there can be multiple orchestrators, particularly during network
partitions. In this case we expect each side of a network partition to
have an orchestrator node provided the network is partitioned for long
enough.

In such a scenario, partition A may try to fail over all of the nodes in
partition B, and partition B may try to fail over all of the nodes in
partition A. It is allowed and acceptable to fail over either partition,
provided that we have 2 or more replicas. As partition A has a quorum,
and partition B does not, the nodes in partition B will be failed over
by the orchestrator of partition A. When the network partition heals,
one of the orchestrator nodes will yield to another. In this case the
orchestrator of partition A will yield to the orchestrator of partition
B.

The orchestrator of partition B will have been attempting to fail over
the nodes in partition A up til this point. Prior to this change, the
safety checks done by auto_failover, such as the preservation of durable
mutations, would be performed before gathering any quorum of nodes.
This, combined with a lengthy (in machine time) quorum gathering timeout
of 15 seconds lead to a scenario in which the orchestrator from
partition B would pass auto_failover checks with:

    1) a view of the nodes in a partitioned state
    2) a config prior to the network partition (as no material change is
       possible without an orchestrator)

In simpler terms, the orchestrator of partition B performed safety
checks without knowledge that the nodes of partition B had already been
failed over. This could then lead to the failover of all of the KV nodes
in the cluster, and 100% data loss.

Correct this issue by taking the quorum and syncing the config before
checking any auto_failover safety checks. Taking the quorum prevents any
material change to the config, syncing the config gets us up to date.
Quorum read transactions were considered for some of the safety checks
that we must make here, but it would be far more work to pass the config
snapshot of the transaction down the failover paths.

Change-Id: I7392e85e29ff2dca3af8d44b6f4f1e221fc5085c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199241
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>