MB-58899: Bump auto_failover gen_server call timeout to 60s
It's possibly (and potentially likely) that we see reconfigure attempts
timeout when auto_failover is processing an auto_failover and is
unresponsive to reconfigure attempts (handle_calls). The main culprit is
likely to be when we don't have a quorum as we use a default timeout of
15s there. The gen_server default timeout is 5s and will cause the REST
request to 500, although it should be processed eventually.

Bump the auto_failover gen_server call timeout to 60s by default (and
make it configurable in case that's every necessary).

Change-Id: Ia58f96ad01389a32554e8fc824df87678c80ede9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198218
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>