MB-59085: add query memoryQuota to pools/default
Add memoryQuota capability to n1ql service, but tie it to queryNodeQuota
because they already have a 'memoryQuota' field in their
json_settings_manager, but it means something completely different for
them (it represents per-query memory usage). This is a bit confusing,
but is handled by overriding the lens to point to another entry (in this
case node-quota).

Change-Id: Id463f091fbe158e3c44919ef86ceaab32b26a8e6
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198678
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>