cluster_tests: Add --collect-logs-after-error param
To avoid having to reproduce a test failure locally to get more info,
we should be able to perform log collection after a failed test.
To avoid potentially adding a lot of unnecessary test time, we only
run a small number of the log collection tasks, and we only run it
immediately prior to shutting down a cluster, rather than after each
testset on the cluster.

Change-Id: Iab107c11d701d6b1758974ccf24886cb1b9a1abd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198629
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>