MB-58879: plain auth hash migration.
- If the current global password hash algorithm/settings are different
from the hash settings for an user, migrate the hashes to use the new
algorithm/settings when user is being authenticated. The feature can be
enabled using the following knob when the entire cluster is on Trinity.

POST "/settings/security/allowHashMigrationDuringAuth" -d "true".

- User password could be concurrently modified while a hash is being
migrated. To prevent such updates from being lost, password changes
are made at a higher priority. The merge function in replicated_storage
therefore now account for the priority of the update (in addition to the
version) before accepting a replicated doc.

Change-Id: Id1e407db250594252fe824131333dcbcb8bf2644
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199665
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>