MB-59134: Fix when failover completion is reported
Since we now janitor a group of buckets together in parallel during
failover,  this fixes the reporting of failover completion such that
it reports it for the entire group of buckets when the group completes
instead of reporting it at the end when all bucket groups have completed.

It also fixes reporting for non membase bucket types to report
completion of them as soon as config as been updated. We don't janitor
non membase buckets, so by definition the failover is complete when
the config has been updated.

Change-Id: I263e9960031859bb7cc1aa0b123941f88ad127c8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198924
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>