MB-58069 Report derived stats in /metrics results
With the introduction of derived stats (stats calculated in prometheus)
the stats are only in prometheus. For the /metrics endpoint we want to
report the derived metrics for backwards compatibility. These metrics
are marked as deprecated so they can be removed in the future after
allowing some amount of time to transition to the replacements.

Change-Id: I17873ce18e23a5dc8f89dbff39df3d0afef66ac6
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199309
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>