MB-58069 Change derived stat names for backwards compatibility
For backwards compatibility (e.g. clients of /metrics) we will report
the following stats:

sys_cpu_utilization_rate is calculated as either:
        irate(sys_cpu_cgroup_seconds_total{mode=`usage`}
  or
        sys_host_cpu_utilization_rate
depending if Couchbase is part of a cgroup.

sys_cpu_user_rate is calculated as either:
        irate(sys_cpu_cgroup_seconds_total{mode=`user`}
  or
        sys_host_cpu_user_rate
depending if Couchbase is part of a cgroup.

sys_cpu_sys_rate is calculated as either:
        irate(sys_cpu_cgroup_seconds_total{mode=`sys`}
  or
        sys_host_cpu_sys_rate
depending if Couchbase is part of a cgroup.

sys_cpu_throttled_rate is calcuated as:
        irate(sys_cpu_cgroup_seconds_total{mode=`throttled`}
if Couchbase is part of a group. Otherwise it is not reported,

sys_cpu_burstd_rate is calcuated as:
        irate(sys_cpu_cgroup_seconds_total{mode=`burst`}
if Couchbase is part of a group. Otherwise it is not reported,

Change-Id: I4ae6597d53f8f3ed4332d299a03e3ee1a888e609
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199485
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>