Make client cert auth modification partially synchronous
Sync guarantees that https server has restarted on local node.
It is not guaranteed that other servers on local node has
restarted though. It is also not guaranteed that https servers
on other cluster nodes have restarted. Nevertheless, the fact
that this command returns only when the settings are applied on
local web server, makes writing some scripts (and tests) a lot
easier.

Change-Id: If6e29cd1da0f2ac23174ef2f0e4d666bb21b05cc
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199486
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>