MB-59322: Relax guardrail thresholds
In order to avoid data ingress being disabled too early, we are
relaxing the thresholds, to allow the opportunity for the issue to be
resolved before any writes are rejected.

The RR% thresholds are changed to 1% and 0.2% for couchstore and magma
respectively. We are giving ourselves a 10x/5x headroom, as we expect
to see quite a number of clusters breaching the recommended minimums.

The max data per node threshold is only increased for couchstore,
rounding 1.6 up to 2. For magma, 16TB is such a large amount of
active-only data to have on a node, so we don't currently see the need
for more headroom.

The disk usage threshold is increased to 96%, leaving enough room that
we should hopefully not see the disk become completely full before
blocking writes.

We are keeping the cores/bucket and quota/collection limits the same,
as these are config-time limits, so no writes would be rejected with
these being enforced.

For the config-time checks of all guardrails, control plane is now
expected to perform the enforcement, with ns_server still making them
(albeit later in the case of the relaxed thresholds noted above) as a
fallback. This covers rebalance checks, bucket updates, and collection
limits. The only checks that control plane is not expected to make are
those which can only be enforced by ns_server, that is those that
disable data ingress.

Change-Id: Ibefa831b43e9588e62c97e19a5d41c34ec783db3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199670
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>