MB-58768: Update use of cluster.nodes in auto_failover_test
When cluster.connected_nodes was introduced, it meant that
cluster.nodes no longer could be relied upon to get a node in the
cluster. auto_failover_test wasn't modified to take that into account
at the same time, so must be fixed, to avoid it failing in future if it
gets executed on a cluster where cluster.nodes[0] is not in the
cluster.

Change-Id: I3bd4fb71ffb458c33aeeb25212d592a580b36fa4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200863
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>