MB-58768: Update use of cluster.nodes in bucket_test
When cluster.connected_nodes was introduced, it meant that
cluster.nodes no longer could be relied upon to get a node in the
cluster. bucket_test wasn't modified to take that into account at the
same time, so must be fixed, to avoid it failing in future if it gets
executed on a cluster where cluster.nodes[0] is not in the cluster.

Change-Id: I3a544c5d2e022762f4bc5475912846ad742dbe6e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200865
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>