MB-58768: Rename Cluster.nodes to Cluster._nodes
It's easy to type cluster.nodes, without remembering that you
specifically need nodes that are in the cluster. By renaming it to
_nodes, it should be much harder to make this mistake.

To get a node that is not in the cluster, the cluster.spare_node() and
cluster.disconnected_nodes() functions should cover most use cases.
This leaves cluster._nodes as only being required when you definitely
need every node in and out of the cluster.

Change-Id: I4d3cb494f52206e1e711805d370f582657b503e2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200936
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>