Restrict parallelism for recovery tests
It is very easy to hit the file descriptor limit on MacOS. This test
is not particularly deterministic. When executed with 1 replica it
will use the "old" mb_map approach. For replica counts greater than 1
it will use the new approach that writes to files and requires file
descriptors to interact with vbmap. With sufficient consecutive tests
hitting the new vbmap approach it is very easy to hit the max file
descriptor for a process on MacOS.

Prior to this change the recovery tests took on average 8s to run
locally. After this change they took on average 11s.

Change-Id: I733b394b7e67b1ec498cd14d959471baafe4ed7b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199886
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>