MB-59263: Provide 'atom' reason cluster balance
Provide an atom code to associate with the "reason" for which the
cluster is not balanced. This originally only included nicer error
messages but will now include an atom code for easier matching on the
control-planes side. Also de-duplicated a lot of the formatted text
which was entirely redundant and instead group by the error type and
allow multiple in the servers or buckets fields.

Example:
...
"balanced": false,
"servicesNeedRebalance": [
    {
      "code": "service_not_balanced",
      "description": "Service needs rebalance.",
      "services": ["kv", "n1ql"],
    }
],
"bucketsNeedRebalance": [
    {
      "code": "num_replicas_changed",
      "description": "Number of replicas for bucket has changed.",
      "buckets": ["default", "another"]
    },
    {
      "code": "servers_not_balanced",
      "description": "Bucket map needs rebalance.",
      "buckets": ["another"]
    }
],
...

Change-Id: I2de21583ac7918403815c15bf40cb3e24b50d7d3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199798
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>