cluster_tests: Handle request errors in smog_check
By asserting that the request to /pools/default or /pools has status
200, we crash cluster_tests entirely if a testset leaves the cluster in
a state where we get a different status code. These unexpected status
codes should just be treated as an inconsistent cluster, and handled
accordingly.
Additionally, if a node was just removed/failed over, it may not be
responsive, so we'll get an exception from the request to /pools.
We should handle this exception, since it is evidence of the cluster
being in an inconsistent state, and so we should show this as a testset
error, rather than just letting cluster_tests crash.

Change-Id: I530ca6875392ea439dc46293de583dd37a6cdbad
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199895
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>