MB-59387: Fetch cpu count directly instead of using ns_doctor
Using ns_doctor to fetch a newly added node's cpu count meant that
there was a race with the first heartbeat from that node, which meant
that in some cases a rebalance would permit adding a node with
insufficient cores as it ignored the node if it had an undefined
cpu count.
While we could simply reject rebalances where we don't yet have the cpu
count for a newly added node, this would not be ideal behaviour.
Instead we should directly call sigar to get the cpu count, and reject
the rebalance if we don't receive a number.

There is a risk here that we will reject a rebalance because of an
error getting the cpu count from sigar, in which case we give a
different error message, specifying that there was a problem getting
the cpu count, rather than just that there are insufficient cpu cores.

Change-Id: Ibf531993c1a0f9182284fc9a14dc1d02f4ef1262
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199917
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>