MB-58863: Permit swap rebalances when disk usage GR breached
A swap rebalance is unlikely to make a disk usage issue worse, so we
should permit the rebalance to go ahead if safe, even if the disk usage
limit has been reached. A later commit may be added to determine if we
actually expect the swap rebalance to be safe to perform.

Change-Id: I9f5ff5aec0cf9bd9750c656f5bb3a55842121254
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199999
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>