Fix GuardRailRestrictionTests.storage_migration_test
While we no longer need to wait for the janitor for the migration to be
permitted, the guardrail still needs to know which servers the bucket
will be on in order to fetch the data size stats for the bucket.
Since we shouldn't be able to hit the data guardrails before the
servers list is populated (realistically speaking), this isn't really a
bug in the code, so it should be sufficient to simply add the
is_servers_populated polling back into the test.

Change-Id: I41533f05e19f68f99e8262acd2abaa72c5185d02
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200066
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>