MB-59130: Reject increase of num_replicas if disk GR breached
If the disk usage guardrail has been breached on a node, we should not
allow increasing the number of replicas for any buckets on the node, as
this would increase the disk usage further (assuming that no additional
nodes are added at the same time).

Ideally we would check if there is space on the disks for the replicas,
irrespective of whether a node has already breached the disk guardrail.
If we decide to do that, it will be in a subsequent change, as the
check would need to be at rebalance time and would be much more
complicated.

Change-Id: Ie362068caeeaaa1b2dfc7f5d51c2893848675156
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200342
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>