MB-58863: Use live stats for disk guardrail checks
In order to get the disk stats for a node that we are adding to the
cluster, we need to be able to get them immediately, without waiting
for a prometheus scrape, so we must instead use ns_disksup directly.
To maintain consistency, we should also make all of our other disk
usage checks be based on ns_disksup rather than prometheus, which means
that those checks are no longer based on a configurable PromQL query.

No longer being based on a PromQL query could lead to inconsistency
with the control plane's checks. The impact it will have is that we may
react to changes in the state of a nodes earlier than the control plane
is able to, which could mean the control plane trying to do something
which we do not permit. In the worst case, all this will mean is that
the server stops the control plane from doing something that we know to
be more unsafe than the control plane realizes. Since we changed the
guardrail thresholds to be much less strict, this is actually very
unlikely to happen, as the control plane is expected to take actions
much earlier than we react.

Another consequence of this is that the corresponding cluster tests can
no longer simulate a disk usage issue by simply changing the promQL.
Instead we must simply set the maximum disk usage to 0, which ensures
that the guardrail kicks in. To avoid completing this rebalance, the
tests send a request to stop the rebalance, then another to perform
another rebalance which should undo any changes already made by the
original rebalance. So that this second rebalance is still permitted,
we must now additionally disable the guardrails as part of the
rebalance cleanup, and reset the settings back to their prior state
again after the second rebalance is completed.

Change-Id: I1229b0c8be005d62e7a8ef78c2875916da5e7b7f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200333
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>