Fix janitor race in tests for data rebalance guardrail checks
Before the janitor runs for a bucket, the servers list for the bucket
is []. This is treated as the bucket being on 0 nodes.
This mean that if the guardrail was already hit before the janitor
runs, then we would permit a rebalance that removed a node, as it would
look like it increased the number of nodes for the bucket.

In reality, this should never happen as we won't get any data in the
bucket before the servers list is populated.
However, the GuardRailRestrictionTests.rebalance_rr_test and
rebalance_data_size_test fake the data size to trigger the
corresponding guardrails, and this could happen before the janitor run
for the bucket, which makes those tests intermittantly fail.
Because of this, we must wait for the server prop to be populated, at
the start of each of these tests.

To do this generically, a "sync" option has been added to
Cluster.create_bucket, making it wait for the bucket to be healthy on
all nodes before returning.

Change-Id: I4ab8c54140c980bb06fa6d784e823a97cde4c5b3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200499
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>