MB-58898: Add missing meck new/unload to test
This lack of an unload call caused issues when merging to master as a
module was already mecked then we called meck:new() for it. The
meck:new() call is not technically required, meck would create the
meck for us (without the passthrough option), but manual meck:new() is
the norm in our code base so add that too.

Change-Id: I4c1d649a76578c42a974e4210588899a61a4cbe7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200584
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>