cluster_run_lib: Use a timeout by default for all requests
By default there is no request timeout for urllib requests, which means
that in certain circumstances (such as a SIGSTOP'd node), cluster_tests
can get stuck forever waiting for a request to complete. To avoid this
issue, we should specify a timeout of 60s for all requests in
cluster_run_lib. We should not expect to ever hit this timeout in
normal operation, as we have already been using a timeout of at most
60s at a higher level for waiting for a node to be up, which is the
case where we are most likely to hit a high timeout.

Change-Id: I880da7770eb7e70d8a1b206be62bf3041f9d7347
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200585
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>