MB-59599: Update membase and memcached buckets config atomically
Currently we are handling config updates for the two bucket types
separately, but there is no reason to. memcached buckets have no maps
so only server list needs an update, so this patch combines the two
update types into a single transaction.

Also in the past we always avoided writing the map again for membase
buckets, if we find an empty map. Now we always write the map, even if
it is empty and the current value is empty. This shields from spurious
config updates being applied in the future from past dead clients.

Change-Id: I13d53530b7f81b293736782f1d2b7523c38194b9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200751
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>