MB-58027: Only request dcp stats for replication connections
We do not need stats for non-replication connections and memcached may
return a substantial amount of stats if there are a lot of connections
or streams. Processing of these stats that we would just discard anyway
adds pointless work and memory usage in both KV and ns_server.

Change-Id: I2c852cc8d8232b80e150258a4e9e00e443d6e1e1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200841
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>