cluster_tests: Get existing connected nodes in any order
When using an existing cluster for cluster tests, if cluster tests has
already run on the cluster, then the first node may no longer be
connected. In order to be able to reuse this cluster for cluster tests,
we now check each node in the provided range of ports to find which
ones are connected.

Change-Id: Id6ca28b4062867e9207d8df5214abc90d2f33d04
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200861
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>