MB-59522: Explicitly handle "good" statuses for takeover stats
KV should always return "calculating-item-count", "backfilling" or
"in-memory" statuses when things are working as they ought to be.
Currently KV will return "does_not_exist" for various other errors, and
we will default to "does_not_exist" for various errors internally in
ns_server too.

Explicitly handle the "good" statuses and bubble the status up the stack
as an error if not. This will let us change the statuses returning by KV
or when we detect other errors, so that we can determine what the error
was without guess work based on log messages.

Change-Id: Id905b5515746a69b8ea0a9c3ef6d0b9d0fafd762
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200933
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>