MB-58768: Update use of cluster.nodes in saml_tests
When cluster.connected_nodes was introduced, it meant that
cluster.nodes no longer could be relied upon to get a node in the
cluster. saml_tests wasn't modified to take that into account at the
same time, so must be fixed, to avoid it failing in future if it gets
executed on a cluster where cluster.nodes[0] is not in the cluster.

Change-Id: Ic784b0f2abfa07253f70bd396bce5fe61195bc0f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200935
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>