MB-59455 terminate all external cbauth connections if
...menelaus_cbauth goes down. This is needed because revrpc connections
survive ns_server restart and therefore it can happen that external
cbauth connection is still alive when users database is erased when
node leaves the cluster.

Downside of this change is that external clients will have to re-connect
of menelaus_cbauth crashes for whatever reason. Which should not
be a problem.

Change-Id: I145d5bb7680220176c42b7d0da677b1c6accac9f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200973
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>