MB-58898: Correct domain usage of leader activities by failover
Failover will currently set the domain of a leader activity to
auto_failover if an auto_failover is being ran. I originally believed
that this would be necessary to nest leader activities, but this is,
in fact, problematic, as in theory any leader activity on a different
domain could proceed in parallel. This behaviour change was not
intentional, correct this by using the default domain in failover
regardless of the type of failover.

Change-Id: I9a417d465c4773e4ce0c79444c1bd1766e30ce9b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201022
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>