MB-58491 Remove use of validator:changeable*
MB-58517 [Mixed Version]: Unable to create collection with maxTTL

The validator:changeable* functions are no longer called to validate
REST parameters. The reason is the value specified in the validation
might be the same as the existing value for the entity...thus
technically it's not being changed.

This fixes a issue where you couldn't create a collection with a maxTTL
value specified if the cluster wasn't yet trinity.

Change-Id: I329ee6a5934da9b34e9798bc0dac40d55c1e8181
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201259
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>