Log bucket map diff consistently with scopes diff
When logging an update to the collections key, we avoid a long log
message for the scopes property by replacing it with a diff, under the
key 'scopes_diff'.
However, for the bucket props 'map' and 'fastForwardMap', we do a
similar diff, but use the original key, which makes the log message
potentially misleading, as it could look like those diffs are the new
values. To avoid this confusion, we should use 'map_diff' and
'fastForwardMap_diff', which also makes it consistent with
'scopes_diff'.

Change-Id: I58785c6a91a6dd17adfe5f371e70c5089529c2bb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201224
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>