MB-59724: Correctly set tick period on auto_failover timeout update
It was accidentally removed in a previous change as I forgot why I
wrote some code the way I did. We should update the tick period whenever
we update the timeout value as the two are intrinsically linked. Add a
unit test to guard this too.

Change-Id: Ifde2da6a6befeb6f82786952193fce5b0a01588e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201227
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>