Use function instead of state match to get auto_failover errors in test
Currently we get the reported errors in a test by matching on the state
record of the auto_failover process. This creates work for anybody
changing it in the future. Add a (test only) function to auto_failover
to return the state that we care about.

Change-Id: I00c8bf4a838100a4b08a7ef36014c8692ba61761
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201235
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>