MB-59791: Evaluate beginning '//' URI correctly
When a URI starts with '//' (or more slashes) it is parsed as though the
following section (until next /) is considered the HOST. This is
problematic because then the path and host are separated, and the host
is completely ignored. This means you can pass a path with an invalid
host section to skip over checks against calling /diag/eval endpoints
from the N1QL curl requests.

There are other counter-measures being taken from their (n1ql's) end as
well. This means if you pass a URI like this now, you will receive a 400
error code and the message "Bad Request". Also includes a "warning"
loglevel message to explain this more in the ns_server.debug.log.

Change-Id: Iec948d215ca4fc76f55d0435a3045b8a934cba7c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201719
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>