MB-59174: Prefer 404 to 405 when URL path is ...
... not existent. Many web servers work in the opposite order, where the
path is evaluated, and if it exists the methods are checked. We have
designed our web server router differently and instead break it down
based on METHOD and then by PATH. The easiest "fix" for this is to
always return 404 in these cases instead of 405 since other web servers
function this way as well.

It is common to use 404 in these situations since some people find 405
to be a security issue that helps attackers map the "valid" endpoints.
Since reordering the way we evaluate METHOD/PATH is non-trivial the
easiest answer is to just return 404 in these cases as well.

Since the RFC doesn't appear to specify what order these should be
evaluated, I think it is perfectly valid for us to function this way.

Change-Id: Ia6439bcee0c88b21a22de5a9f69d746569edc0e7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201728
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>