MB-59884: Don't repeatedly set data ingress
The code to set the data ingress status on bucket warmup was also
running every time janitor performed cleanup. This meant an unncessary
memcached call to every bucket on every node every 5s when guardrails
are enabled.

To avoid this, we must instead make the check at the point where we
know that the bucket is not yet warmed up, but just prior to enabling
traffic. This means that we need to pass the data ingress status within
the mark_warmed call.

Change-Id: I755c831cc33e321cf73f48970c4ecc036758e74b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201773
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>