MB-59908: Don't crash guardrail_enforcer when node goes down
If a node is down when the guardrail enforcer tries to set the data
ingress status, it will try to call the down node and crash. It should
catch the error, and treat it like other errors that could occur at
that point, by attempting to set the data ingress status again the next
time it checks.

Change-Id: I4ebaaf5abdd6153249c5d4f5ce0a2e2b5f270c6f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201892
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>