MB-59085: Ensure n1ql default quota is always 0
Modify default remaining quota calculation to always return 0 instead of
using the default formula. This is especially important because the
setDefaultMemQuotas setting that will use this to calculate the default.
This causes the value to end up being 853 which is not what we want the
default to be.

Change-Id: Id66f6bc52e9a6365ff1e35863a94a54670f935af
Reviewed-on: https://review.couchbase.org/c/ns_server/+/202336
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>