MB-59772 Re-add code lost when change-id changed.
I first had this change:
         https://review.couchbase.org/c/ns_server/+/202000
under review. I changed the change-id to match a dependancy with a
differ repo. I then made a change based on the code review. And that was
approved.

I then abandoned the 2nd change-id and merged with the first. I lost the
changes made based on the code review. Note: on my laptop both set of
changes were in the same repo.

This change adds back the changes made based on the code review.

Lessons learned: if the change-id is changed (e.g. to patch a
dependancy) then you should continue to use the new change-id. This is
especially true if any changes are made to the repo.

Change-Id: I9937d837a403f15b1b50a8de35b4a1439aadb164
Reviewed-on: https://review.couchbase.org/c/ns_server/+/202383
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>