MB-58932: Support rebalance report in rebalance_analyser
In addition to generating a rebalance plot from master_events.log, it
now can instead take a rebalance report, and produce a similar plot.

For example:

./scripts/rebalance_analyser.py -r .../rebalance_report.json -b default

See on the ticket for an example comparing the two plots.
Typically, there will only be negligible differences in timestamps, as
the report uses a less precise time format. This can affect the order
that the vbuckets are displayed in, so some care is required to compare
the two plots.
In more unusual scenarios, the difference can be quite extreme, as
the report depends on the rebalance_observer, which can get stuck,
causing the report to be much less useful than the master_events.log.
Despite this, it is useful to support both, as we have previously seen
cases where we have a rebalance report but not a master_events.log.

Change-Id: I0322de41f42ffa7da4313778dddb0279850c76ac
Reviewed-on: https://review.couchbase.org/c/ns_server/+/202512
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>