MB-58932: Rename scripts dirs and files for python imports
In order to import python code in scripts/jq/master-events directory
(without importlib), it needs to have a snake case name.
The script plot-bucket-rebalance.py is also renamed to snake case, so
that it is a valid python module.

This is needed for cleanly calling the code from the rebalance
summariser script in a subsequent change.

Change-Id: Icf16a16d7508c506e73bce110edc34a3bb9bd35c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/202511
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>