MB-60026: Test argon2id hash params before applying them
Argon2id hash calculation is a bit special. It is executed in a nif
and is supposed to take a lot of cpu and memory by design (that's
the point of that hash).
Erlang can't interupt nifs, so if the hash parameters are too brutal
it will lead to a nif running forever taking 1 dirty scheduler with
it (+ 1 full cpu core, + probably a lot of memory). Running multiple
such hash calculations in parallel will 100% lead to a disaster
(especially if such parameter settings get to ns_config).
In order to prevent this we should test parameters before applying
them, but we can't test ANY kind of parameters (because we can't
stop nifs), so we run a heuristic check first (to make sure
the parameters are sane, and hash calculation will not take forever),
and only then we run the real hash calculation.

Heuristic check: (Time * Memory) < 20_000_000_000

Change-Id: I7a8567e3cf90f63e0f8895ef235db8b1d72bb829
Reviewed-on: https://review.couchbase.org/c/ns_server/+/203229
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>