MB-58718: Add Buckets requirement
Most TestSets utilise buckets in some way or another. This typically
requires adding a bucket creation step in setup and/or bucket deletion
in teardown. To simplify this, the expected set of buckets should be
a Requirement, and the test framework should handle satisfing this
requirement before/after the testset runs.

This means that testsets shouldn't need to delete any buckets that they
create while running, as these will be deleted afterwards, if necessary
to satisfy requirements. It also means that any bucket that would
otherwise be specified in the TestSet's setup should instead just be
included in the Buckets requirement.

Currently, no attempt is made to reuse any buckets if the set of
buckets does not exactly match the specified requirements, instead all
existing buckets are deleted and all required buckets are subsequently
created. It should be possible to optimise this to modify existing
buckets that can satisfy the requirements, or at least keep any buckets
that already satisfy the requirements, but this optimisation is
unlikely to have a particularly significant impact so it is not yet
being pursued.

The Buckets requirement takes a list of dicts, where each dict is the
set of props to pass to the bucket creation endpoint. Invalid props
are not currently detected by the requirement, so they will instead
cause the Requirement.make_met step to fail, so care should be taken to
ensure that the props will always be valid for any cluster satisfying
the set of requirements specified.

When checking if the requirement is satisfied, certain props may not
automatically be checked correctly, as the bucket info payload doesn't
follow the exact same format as the creation payload. If one of these
props is required, explicit handling will need to be added in
Buckets.get_prop.

Change-Id: Ibca3b62498a410d1250009dc8c1b472099553341
Reviewed-on: https://review.couchbase.org/c/ns_server/+/205224
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>